Get XCMSnExp object in one step from structured folder path for xcms 3.

Get XCMSnExp object in one step from structured folder path for xcms 3.

getdata2(path, index = F, snames = NULL, sclass = NULL,
  phenoData = NULL, BPPARAM = BiocParallel::SnowParam(), mode = "onDisk",
  ppp = xcms::CentWaveParam(ppm = 5, peakwidth = c(5, 25), prefilter = c(3,
  5000)), rtp = xcms::PeakGroupsParam(minFraction = 0.67),
  gpp = xcms::PeakDensityParam(sampleGroups = 1, minFraction = 0.67, bw = 2,
  binSize = 0.025), fpp = xcms::FillChromPeaksParam())

Arguments

path
the path to your data

index
the index of the files

snames
sample names. By default the file name without extension is used

sclass
sample classes.

phenoData
data.frame or NAnnotatedDataFrame defining the sample names and classes and other sample related properties. If not provided, the argument sclass or the subdirectories in which the samples are stored will be used to specify sample grouping.

BPPARAM
used for BiocParallel package

mode
'inMemory' or 'onDisk' see `?MSnbase::readMSData` for details, default 'onDisk'

ppp
parameters for peaks picking, e.g. xcms::CentWaveParam()

rtp
parameters for retention time correction, e.g. xcms::ObiwarpParam()

gpp
parameters for peaks grouping, e.g. xcms::PeakDensityParam()

fpp
parameters for peaks filling, e.g. xcms::FillChromPeaksParam(), PeakGroupsParam()

Value

a XCMSnExp object with processed data

Details

This is a wrap function for metabolomics data process for xcms 3.

See also

getdata,getupload2, getmzrt2