Get the features from anova, with p value, q value, rsd and power restriction

Get the features from anova, with p value, q value, rsd and power restriction

getfeaturesanova(list, power = 0.8, pt = 0.05, qt = 0.05, n = 3,
  ng = 3, rsdcf = 100, inscf = 5, imputation = "l")

Arguments

list
list with data as peaks list, mz, rt and group information (more than two groups)

power
defined power

pt
p value threshold

qt
q value threshold, BH adjust

n
sample numbers in one group

ng
group numbers

rsdcf
the rsd cutoff of all peaks in all group

inscf
Log intensity cutoff for peaks across samples. If any peaks show a intensity higher than the cutoff in any samples, this peaks would not be filtered. default 5

imputation
parameters for `getimputation` function method

Value

dataframe with peaks fit the setting above