Get the mzrt profile and group information for batch correction and plot as a list directly from path with default setting

Get the mzrt profile and group information for batch correction and plot as a list directly from path with default setting

getmr(path, index = F, BPPARAM = BiocParallel::SnowParam(),
  pmethod = "hplcorbitrap", minfrac = 0.67, ...)

Arguments

path
the path to your data

index
the index of the files

BPPARAM
used for BiocParallel package

pmethod
parameters used for different instrumentals such as 'hplcorbitrap', 'uplcorbitrap', 'hplcqtof', 'hplchqtof', 'uplcqtof', 'uplchqtof'. The parameters were from the references

minfrac
minimum fraction of samples necessary in at least one of the sample groups for it to be a valid group, default 0.67

...
arguments for xcmsSet function

Value

list with rtmz profile and group infomation

See also

getdata,getupload, getmzrt, getdoe

Examples

# NOT RUN {
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
list <- getmr(cdfpath, pmethod = ' ')
# }