Get the report for technique replicates.

Get the report for technique replicates.

gettechrep(xset, method = "medret", intensity = "into", file = NULL,
  rsdcf = 30, inscf = 1000)

Arguments

xset
the xcmsset object which for all of your technique replicates for one sample

method
parameter for groupval function

intensity
parameter for groupval function

file
file name for further annotation, default NULL

rsdcf
rsd cutoff for peaks, default 30

inscf
intensity cutoff for peaks, default 1000

Value

dataframe with mean, standard deviation and RSD for those technique replicates combined with raw data