plot EIC and boxplot for all peaks and return diffreport

plot EIC and boxplot for all peaks and return diffreport

plote(xset, name = "test", test = "t", nonpara = "n", ...)

Arguments

xset
xcmsset object

name
filebase of the sub dir

test
't' means two-sample welch t-test, 't.equalvar' means two-sample welch t-test with equal variance, 'wilcoxon' means rank sum wilcoxon test, 'f' means F-test, 'pairt' means paired t test, 'blockf' means Two-way analysis of variance, default 't'

nonpara
'y' means using nonparametric ranked data, 'n' means original data

...
other parameters for `diffreport`

Value

diffreport and pdf figure for EIC and boxplot

Examples

if (FALSE) {
library(faahKO)
cdfpath <- system.file('cdf', package = 'faahKO')
xset <- getdata(cdfpath, pmethod = ' ')
plote(xset)
}