Surrogate variable analysis(SVA) to correct the unknown batch effects

Surrogate variable analysis(SVA) to correct the unknown batch effects

Usage

svacor(xset, lv = NULL, method = "medret", intensity = "into")

Arguments

xset

xcmsset object

lv

group information

method

parameter for groupval function

intensity

parameter for groupval function

Value

list object with various components such raw data, corrected data, signal part, random errors part, batch part, p-values, q-values, mass, rt, Posterior Probabilities of Surrogate variables and Posterior Probabilities of Mod. If no surrogate variable found, corresponding part would miss.

Details

this is used for reviesed version of SVA to correct the unknown batch effects

See also

Examples

if (FALSE) { # \dontrun{
library(faahKO)
cdfpath <- system.file("cdf", package = "faahKO")
cdffiles <- list.files(cdfpath, recursive = TRUE, full.names = TRUE)
xset <- xcmsSet(cdffiles)
xset <- group(xset)
xset2 <- retcor(xset, family = "symmetric", plottype = "mdevden")
xset2 <- group(xset2, bw = 10)
xset3 <- fillPeaks(xset2)
df <- svacor(xset3)
} # }